fix: address review finding — eliminate data round-trip anti-pattern in generate-address-block (P2)
Extract block data computation from block.ts into server/services/block-data.ts.
The generate-address-block endpoint now fetches block data server-side using
the validated lat/lng/radius instead of trusting client-supplied blockMetrics.
Frontend updated to send radiusMiles instead of full blockMetrics object.
