+++ src/api/client.ts
@@ -79,13 +79,13 @@ export function generateAddressBlockReport(
   lat: number,
   lng: number,
   communityName: string,
-  blockMetrics: BlockMetrics,
+  radiusMiles: number,
   language: string,
   communityMetrics?: { resolutionRate: number; totalRequests: number } | null,
 ): Promise<CommunityReport> {
   return fetchJSON(`${BASE}/report/generate-address-block`, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ address, lat, lng, communityName, blockMetrics, language, communityMetrics }),
+    body: JSON.stringify({ address, lat, lng, communityName, radiusMiles, language, communityMetrics }),
   });
 }
